pyproject.toml
[project]
name = "mamba-memory"
version = "0.1.0"
description = "AI Agent Cognitive Memory Engine — SSM-inspired selective memory with decay, gating, and state evolution"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "MambaMemory Team" }]
keywords = ["ai", "memory", "agent", "ssm", "mamba", "cognitive", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "numpy>=1.26",
    "hnswlib>=0.8",
    "pydantic>=2.6",
]

[project.optional-dependencies]
server = [
    "mcp>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
]
google = [
    "google-genai>=1.0",
]
openai = [
    "openai>=1.40",
]
local = [
    "sentence-transformers>=3.0",
]
llm = [
    "openai>=1.40",
    "anthropic>=0.30",
]
all = [
    "mamba-memory[server,google,openai,local,llm]",
]
dev = [
    "mamba-memory[server,google]",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "anyio>=4.0",
]

[project.urls]
Homepage = "https://github.com/mamba-memory/mamba-memory"
Documentation = "https://github.com/mamba-memory/mamba-memory#readme"
Repository = "https://github.com/mamba-memory/mamba-memory"
Issues = "https://github.com/mamba-memory/mamba-memory/issues"

[project.scripts]
mamba-memory = "mamba_memory.cli:main"

[project.entry-points."mcp.servers"]
mamba-memory = "mamba_memory.server.mcp.server:serve"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mamba_memory"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
